OCR/final_working_interface/jl_interface/requirements.txt
Django==5.0.7
Pillow
pytesseract
python-dotenv
openai
